% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_daily_meteo}
\alias{check_daily_meteo}
\title{Check if meteo data are daily or hourly}
\usage{
check_daily_meteo(self)
}
\arguments{
\item{self}{A \code{user_data} R6 object containing a \code{meteo} data.frame with a column named \code{meteo_datetime}.}
}
\value{
A logical value:
\code{TRUE} — if the dataset represents daily meteo data.
\code{FALSE} — if the dataset contains hourly (or sub-daily) timestamps.
}
\description{
This function analyzes the \code{meteo_datetime} column from the \code{meteo} slot
of the \code{user_data} object and determines if the data are daily or hourly observations.
It is mainly used by \code{\link[=plot_meteo]{plot_meteo()}} to ensure that daily data are plotted correctly.
}
