% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{check_heteroscedasticity}
\alias{check_heteroscedasticity}
\title{Bartlett on residuals}
\usage{
check_heteroscedasticity(model, alpha = 0.05)
}
\arguments{
\item{model}{a linear model (lm) object}

\item{alpha}{alpha value for test}
}
\value{
pvalue of Bartlett's test
}
\description{
This function test homogeneity of variances test thanks to Bartlett's test on residuals
}
