% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{check_normality}
\alias{check_normality}
\title{Residual normality test}
\usage{
check_normality(model, alpha = 0.05)
}
\arguments{
\item{model}{a linear model (lm) object}

\item{alpha}{alpha value for test}
}
\value{
pvalue of Shapiro-Wilk's test
}
\description{
This function test residual normality test thanks to Shapiro-Wilk's test
}
