% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_plotid_diff}
\alias{check_plotid_diff}
\title{Check differences in plot_id}
\usage{
check_plotid_diff(df_as_ref, df_to_test)
}
\arguments{
\item{df_as_ref}{a reference dataframe with a plot_id col}

\item{df_to_test}{a dataframe to test with a plot_id col}
}
\value{
TRUE if differences are detected (NA are ignored)
}
\description{
Check differences between plot_id values between a reference dataframe (typically the plot_desc sheet) and an other dataframe (typically an observation dataset)
}
