% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pom_csv}
\alias{check_pom_csv}
\title{Check for required columns in a data frame}
\usage{
check_pom_csv(filepath, required = c("DATE", "PLUIE"))
}
\arguments{
\item{filepath}{Path to the .csv file to check.}

\item{required}{A character vector of column names that must be present in \code{df}.
Default is \code{c("DATE", "PLUIE")}.}
}
\value{
Returns \code{TRUE} if the csv file is compliant (csv and all required columns presents), else stops execution
with an error message.
}
\description{
This function verifies that a given data frame is compliant to a POM export, that means
it contains all required columns.
If any required column is missing, it stops execution and provides a clear
error message listing missing and present columns. If all required columns
are present, it returns TRUE and prints a confirmation message.
}
\details{
\itemize{
\item The function first checks if \code{filepath} is indeed a csv file.
\item It then compares the names of \code{filepath} to the \code{required} columns.
\item If any columns are missing, an informative error is thrown listing missing
and available columns.
\item If all required columns are present, a message is displayed confirming their presence.
}
}
