% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOut.R
\name{check_standard_file}
\alias{check_standard_file}
\title{Check if the file provided is a xlsx standardised file}
\usage{
check_standard_file(trial_file = NULL)
}
\arguments{
\item{trial_file}{a file path}
}
\value{
TRUE or FALSE
}
\description{
Check if the file provided is a xlsx standardised file
}
\examples{
check_standard_file(trial_file = system.file("extdata","standard_exemple.xlsx",package="startbox"))
}
