% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardise.R
\name{check_topvigne_csv}
\alias{check_topvigne_csv}
\title{Check if a CSV file is in TopVigne format}
\usage{
check_topvigne_csv(filepath)
}
\arguments{
\item{filepath}{Path to the .csv file to check.}
}
\value{
Logical. Returns TRUE if the file is valid. Stops execution with an error message otherwise.
}
\description{
This function reads a .csv file and checks whether it contains the minimum
required columns to be considered a valid TopVigne file.
}
