% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_xptrtcode_diff}
\alias{check_xptrtcode_diff}
\title{Check differences in xp_trt_code}
\usage{
check_xptrtcode_diff(df_as_ref, df_to_test)
}
\arguments{
\item{df_as_ref}{a reference dataframe with a xp_trt_code col}

\item{df_to_test}{a dataframe to test with a xp_trt_code col}
}
\value{
TRUE if differences are detected (NA are ignored)
}
\description{
Check differences between xp_trt_code values between a reference dataframe (typically moda_desc)
and an other dataframe (typically an observation dataset).
Most of the time xp_trt_code are not saved in raw observation datasets :
observation dataset should then be merged before with plot_desc, using merge_data_plotdesc()
}
