% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_vars}
\alias{find_vars}
\title{Look star variables in a dataframe}
\usage{
find_vars(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a vector of variables found in the dataframe that are also in the dictionary
}
\description{
Look star variables in a dataframe
}
