% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metrics.R
\name{frequency}
\alias{frequency}
\title{calcuate frequency in percent of values >0from a numeric vector}
\usage{
frequency(vecteur)
}
\arguments{
\item{vecteur}{}
}
\value{
a numeric value of frequency, in percent
}
\description{
calcuate frequency in percent of values >0from a numeric vector
}
\examples{
vec <- c(1, 2, 3, 0, 5, 6, 0)
frequency(vec)
}
