% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getdataframe}
\alias{getdataframe}
\title{Extract dataframe from an user_data object by his name (looking in obs_data and prepared_data)}
\usage{
getdataframe(self = NULL, dfname = NULL)
}
\arguments{
\item{self}{an instance of the \code{UserData} R6 class containing user data.}

\item{dfname}{a name to found (character)}
}
\value{
a dataframe if found, NULL if not
}
\description{
Extract dataframe from an user_data object by his name (looking in obs_data and prepared_data)
}
