% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{harmonize_all_obs_data}
\alias{harmonize_all_obs_data}
\title{Harmonize All Observation Data Tables}
\usage{
harmonize_all_obs_data(obs_data)
}
\arguments{
\item{obs_data}{A list of observation datasets.}
}
\value{
A list of harmonized observation datasets.
}
\description{
Harmonizes all observation datasets using harmonize_column_types().
}
