% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{harmonize_column_types}
\alias{harmonize_column_types}
\title{Harmonize Column Types}
\usage{
harmonize_column_types(df, dictionary = NULL)
}
\arguments{
\item{df}{The dataframe to harmonize.}

\item{dictionary}{(optional) a data.frame with a least 2 cols 'nom' and 'Rclass'}
}
\value{
A dataframe with harmonized column types.
}
\description{
Harmonizes the types of columns in a dataframe according to a type mapping.
}
