% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge.R
\name{harmonize_plot_id_format}
\alias{harmonize_plot_id_format}
\title{Harmonize plot_id format to match placette (e.g. 10A → A10)}
\usage{
harmonize_plot_id_format(pid)
}
\arguments{
\item{pid}{vector of plot_id (character)}
}
\value{
harmonized plot_id (e.g. A10, TNT3, etc.)
}
\description{
Harmonize plot_id format to match placette (e.g. 10A → A10)
}
