% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ht_efficacy}
\alias{ht_efficacy}
\title{Calcul efficacy according to the Henderson and Tilton formula}
\usage{
ht_efficacy(Ta, Tb, Ca, Cb)
}
\arguments{
\item{Ta}{number of mites after treatment (numeric)}

\item{Tb}{number of mites before treatment (numeric)}

\item{Ca}{number of mites after treatment in the check plot (TNT) (numeric)}

\item{Cb}{number of mites before treatment in the check plot (TNT) (numeric)}
}
\value{
a vector of calculated efficacy
}
\description{
This function calculate the treatment efficacy according to the Henderson and Tilton formula (1955).
#' For non-uniform populations, the Henderson and Tilton (1955) formula is suited which allows us to judge differences in the
population development. In this formula the mortality ratio is corrected on the control mortality
Eff = (1 - (T_a * C_b) / (T_b * C_a)) * 100
with :
Tb number of mites before treatment
Ta  number of mites after treatment
Cb number of mites before treatment in the check plot (TNT)
Ca number of mites after treatment in the check plot (TNT)
Efficacy is expressed in percentage.

Reference : Henderson, C.F. and E. W. Tilton. 1955. Tests with acaricides against the brow wheat mite, J. Econ. Entomol. 48:157-161.
See also : Abbott (1925) formula
}
\examples{
# ht_efficacy(Ta=5,Tb=15,Ca=20,Cb=15)

}
