% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_excel.R
\name{import_data_sheets}
\alias{import_data_sheets}
\title{Import "data_*" sheets from an Excel file into a \code{user_data} object}
\usage{
import_data_sheets(self, filepath)
}
\arguments{
\item{self}{The \code{user_data} R6 object}

\item{filepath}{Path to the Excel file to import}
}
\value{
Updates \code{self$obs_data} with the newly imported sheets
}
\description{
Import "data_*" sheets from an Excel file into a \code{user_data} object
}
