% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOut.R
\name{import_topvigne_csv}
\alias{import_topvigne_csv}
\title{Import topvigne observational data from a Topvigne csv file}
\usage{
import_topvigne_csv(self, filepath, obs_name = NULL, overwrite = FALSE)
}
\arguments{
\item{self}{An \code{user_data} R6 object in which to import the meteo}

\item{filepath}{File path of the csv file to be imported. This file must be a csv file exported from TopVigne (IFV app)}

\item{obs_name}{the name of the observation to add}

\item{overwrite}{Logical. If TRUE, replaces an existing entry with the same name.}
}
\value{
Updates \code{self$obs_data} with the imported data, update also the log. Returns \code{invisible(self$obs_data)}.
}
\description{
This function imports meteorological data from a CSV file exported from the Topvigne App.
The data are standardized before being added to \code{self$obsdata}.
The function also updates the internal log (traceability).
}
\examples{
\dontrun{
# Import new topvigne data into an existing user_data object
file_path <- system.file("extdata","maladie_teisso_2024_11h12min_F1_17.06.csv",package="startbox")
import_topvigne_csv(self = mydata, filepath = file_path, obs_name = "F1")
}

}
