% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{incidence}
\alias{incidence}
\title{calculation of disease incidence}
\usage{
incidence(vector, threshold = 0)
}
\arguments{
\item{vector}{a numeric vector of disease severity observations in percentage}

\item{threshold}{threshold to calculate incidence, zero by default}
}
\value{
a numeric value of incidence, in percent
}
\description{
This function calculates the frequency of the disease in the observed population,
i.e. the number of sampled units (leaves, bunches, plants, etc.) infected expressed
as a percentage of the total number of units observed.
}
\examples{
vec <- c(1, 2, 3, 0, 5, 6, 0)
incidence(vec)
}
