% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metrics.R
\name{intensity}
\alias{intensity}
\title{calcuate intensity (mean of value)}
\usage{
intensity(vecteur)
}
\arguments{
\item{vecteur}{}
}
\value{
a numeric value of intensity
}
\description{
calcuate intensity (mean of value)
}
\examples{
vec <- c(100, 20, 30, 0, 5, 0, 0)
intensity(vec)
}
