% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOut.R
\name{load_default_dictionary}
\alias{load_default_dictionary}
\title{load default dictionary}
\usage{
load_default_dictionary(self)
}
\arguments{
\item{self}{A \code{user_data} R6 object with a \code{traceability} data.frame and an \code{excel_data_trial} path.}
}
\value{
no return value, user_data dictionary is updated
}
\description{
load default dictionary
}
