% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOut.R
\name{load_excel_file}
\alias{load_excel_file}
\title{Load data from an standardized Excel file into R environment}
\usage{
load_excel_file(trial_file = NULL, name = NULL)
}
\arguments{
\item{trial_file}{path to data excel model}

\item{name}{a name of the experiment to be stored}
}
\value{
load data into a new user_data object in the R environment
}
\description{
Load data from an standardized Excel file into R environment
}
\examples{
mydata <- load_excel_file()
}
