% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOut.R
\name{load_meteo}
\alias{load_meteo}
\title{Load meteo data from Excel sheet}
\usage{
load_meteo(self)
}
\arguments{
\item{self}{A \code{user_data} R6 object containing the \code{trial_file} attribute,}
}
\value{
Invisibly returns the loaded data.frame stored in \code{self$meteo}.
}
\description{
This function reads the \code{meteo} sheet from the Excel file associated with a \code{user_data} and stores it in the slot \code{self$meteo}.
If the sheet is not found, \code{self$meteo} is set to \code{NULL}.
}
