% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{merge_data_plotdesc}
\alias{merge_data_plotdesc}
\title{Merged a dataset with plot description}
\usage{
merge_data_plotdesc(self, df, flex = F)
}
\arguments{
\item{self}{An instance of the \code{UserData} R6 class containing metadata$plot_desc.}

\item{df}{a dataframe to join.}

\item{flex}{boolean. if TRUE, 10A A1O or 10a are considered as the same plot_id}
}
\value{
A \code{data.frame} with observation data enriched by plot and treatment metadata.
If no observation data is available, returns only the merged plot + modality metadata.
If required metadata is missing, returns \code{NULL} with a message.
}
\description{
Merges observation data with metadata sheets "placette" if available.
The function joins the 2 dataframes using the \code{plot_id} column.
}
\details{
\itemize{
\item Assumes that \code{self$metadata$plot_desc} contains \code{plot_id}.
}
}
