% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{merge_data_xpdesc}
\alias{merge_data_xpdesc}
\title{Merged a dataset with experimental treatment description}
\usage{
merge_data_xpdesc(self, df)
}
\arguments{
\item{self}{An instance of the \code{UserData} R6 class containing metadata$moda_desc}

\item{df}{a dataframe to join.}
}
\value{
A \code{data.frame} with observation data enriched by ptreatment metadata.
If required metadata is missing, returns \code{NULL} with a message.
}
\description{
Merges observation data with metadata sheets "modalite" if available.
The function joins the 2 dataframes using the \code{xp_trt_code} column.
}
\details{
\itemize{
\item Assumes that \code{self$metadata$moda_desc} contains \code{xp_trt_code}.
}
}
