% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_factor_levels_numeric_last}
\alias{order_factor_levels_numeric_last}
\title{Ordering factors}
\usage{
order_factor_levels_numeric_last(f)
}
\arguments{
\item{f}{a vector of type character or factor to be ordered}
}
\value{
a vector of factors with the numeric one first
}
\description{
Ordering factors
}
