% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpbar2}
\alias{plot_xpbar2}
\title{Barplot summary for an experiment}
\usage{
plot_xpbar2(
  data2plot,
  xcol = "xp_trt_code",
  ycol = "value",
  fillcol = NULL,
  scale = NULL,
  bar_color = NULL,
  bar_width = 0.7,
  border_tnt = TRUE,
  show_errorbar = TRUE,
  ...
)
}
\arguments{
\item{data2plot}{a dataframe, the first col is used for x axis, col value for y axis and col type for fill.}

\item{xcol}{col to be used as factor for x avis}

\item{ycol}{col to be used for y axis}

\item{fillcol}{col to be used for fill}

\item{scale}{Upper limit of the y axis. If NULL, automatic adjustment.}

\item{bar_color}{Bar fill colour.}

\item{bar_width}{bar width}

\item{border_tnt}{if true, the border color is set as red for TNT}

\item{...}{other parameters for labs (title, x, y,fill)}
}
\value{
a barplot
}
\description{
This function automatically generates a bar graph from observation or experimental data.
}
\examples{
# Exemple d'utilisation
plot_xpbar(my_data)
}
