% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpbox}
\alias{plot_xpbox}
\title{Boxplot résumé pour une expérimentation}
\usage{
plot_xpbox(data, titre = NULL, echelle = NULL, caption = "IFV+")
}
\arguments{
\item{data}{Données brutes contenant au moins plot_id et une variable "_PC".}

\item{titre}{Titre principal du graphique (optionnel).}

\item{echelle}{Limite supérieure de l'axe y (optionnel, sinon automatique).}

\item{caption}{A string used as the caption text displayed at the bottom of the plot. Default is "IFV+".}
}
\value{
Un objet \code{ggplot2}.
}
\description{
Cette fonction génère automatiquement un graphique en boîte à moustaches
(boxplot) à partir de données expérimentales brutes (parcelle, observation, maladie).
}
\examples{
plot_xpbox(my_data)
}
