% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpheat}
\alias{plot_xpheat}
\title{Plot a Heatmap of Experimental Data}
\usage{
plot_xpheat(
  self,
  stats = NULL,
  calculation_choices = NULL,
  max_y = NULL,
  flip_orientation = FALSE,
  yx_ratio = 0.3,
  resids = FALSE,
  midpoint = NULL,
  short_names = FALSE,
  ...
)
}
\arguments{
\item{self}{An instance of the \code{UserData} R6 class}

\item{stats}{character, the name of a stats list inside self$stats.}

\item{calculation_choices}{vector of character, choice(s) of calculation to plot. For example c("FA UN_BER_PC","IA UN_BER_PC")}

\item{max_y}{(optional) Numeric. Maximum value for the fill color scale. If NULL, an automatic scale is calculated.}

\item{flip_orientation}{TRUE or FALSE to control plot orientation. Default is FALSE. see ?ggplot2::coord_flip()}

\item{yx_ratio}{specified ratio between the physical representation of data units on the axes. see ?ggplot2::coord_fixed()}

\item{resids}{Logical. If TRUE, residuals are calculated from the linear model and plotted}

\item{midpoint}{numeric, value of midpoint for fill_scale. Automatically set to 0 if resids = T or more than one calculation}

\item{short_names}{to have shorter names for fill removing the last part of the string after the last space occurence}

\item{...}{other parameters for labs (title, x, y,fill, caption). see ?ggplot2::labs()}
}
\value{
A ggplot2 heatmap object.
}
\description{
Generates a heatmap based on experimental observation data,
using \code{plot_x} and \code{plot_y} coordinates and coloring according to a selected disease variable (_PC columns).
}
\details{
\itemize{
\item Aggregates values by \code{plot_id}, taking the mean when multiple observations exist.
\item Useful for visualizing spatial patterns of disease in experimental trials.
}
}
