% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prpnames}
\alias{prpnames}
\title{Names of data observations available in an user_data objetc}
\usage{
prpnames(self)
}
\arguments{
\item{self}{#' @param self an instance of the \code{UserData} R6 class}
}
\value{
a vector of names of prepared data
}
\description{
Names of data observations available in an user_data objetc
}
