% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{qq_residual_plot}
\alias{qq_residual_plot}
\title{Q-Q plot of residuals}
\usage{
qq_residual_plot(model)
}
\arguments{
\item{model}{a linear model (lm) object}
}
\value{
qq plot of residuals
}
\description{
Function to return a Q-Q plot of residuals for a linear model
}
\examples{
# Example usage:
fit <- lm(mpg ~ wt + hp, data = mtcars)
qq_residual_plot(fit)
}
