% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{remove_block_code}
\alias{remove_block_code}
\title{extract experimental traitement code from plot_id}
\usage{
remove_block_code(pid, blocks, separator = NULL)
}
\arguments{
\item{pid}{a vector of character}

\item{blocks}{a vector of block codes (could be a pb if blocks code are a mix of letters and numbers...)}

\item{separator}{a character that separate block code from the other part of the string, for example "_"}
}
\value{
the pid vector without blocks codes
}
\description{
This function remove the block code from a vector of class character.
It works well when block code are letters and experimental treatment code are numbers or vice versa.
TNT are directly recognized as TNT
}
\examples{

blocks=c("A","B","C","D")
pid=c("1A","1B","TNT3","2A","2D")
remove_block_code(pid=pid,blocks=blocks)
}
