% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{remove_code}
\alias{remove_code}
\title{remove specific part of a character string}
\usage{
remove_code(pid, to_remove, separator = NULL)
}
\arguments{
\item{pid}{a vector of character}

\item{to_remove}{a vector of codes to be removed from the string}

\item{separator}{a character that separate parts of the string, for example "_"}
}
\value{
a character vector of same lentgh as pid, without the "to_remove" part
}
\description{
This function remove a specific string from a vector of class character.
It works well when block code are letters and experimental treatment code are numbers or vice versa.
TNT are directly recognized as TNT
}
\examples{

blocks=c("A","B","C","D")
pid=c("1A","1B","TNT3","2A","2D")
remove_code(pid=pid,to_remove=blocks)
}
