% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{remove_string_pairwise}
\alias{remove_string_pairwise}
\title{Remove string in a character vector}
\usage{
remove_string_pairwise(vec, pattern, separator = NULL)
}
\arguments{
\item{vec}{a character vector}

\item{pattern}{a character vector}

\item{separator}{a character that separate parts of the string, for example "_"}
}
\value{
a character vector
}
\description{
This function takes two character vectors of the same length and, for each position, removes the string in pattern\link{i} from vec\link{i}, ignoring case.
If either element at position i equals "TNT" (case-insensitive), it returns "TNT" for that position instead.
}
\examples{
plot_id <- c("1a","2A","1B","tnt1","4c")
plot_block <- c("A","A","B","1","C")
remove_string_pairwise(vec = plot_id, pattern = plot_block)
}
