% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_data.R
\name{resume_data}
\alias{resume_data}
\title{Resume data}
\usage{
resume_data(
  data,
  var_col,
  group_cols,
  funs = list(intensite = mean, frequence = frequency),
  code_tnt = "TNT",
  df_tnt = NULL
)
}
\arguments{
\item{data}{a dataframe to resume}

\item{var_col}{character, the colname of the variable to plot}

\item{group_cols}{the colname of the group}

\item{funs}{statistic to plot, a vector of one or two statistic. by default c("mean","frequency")}

\item{code_tnt}{a string to identify in TNT in the row of the dataframe by default "TNT"}

\item{df_tnt}{a dataframe to associate tnt_id to bloc_id or plot_id. tnt_id is the plot_id of TNT and MUST be a col of df_tnt}
}
\value{
a dataframe ready for plotting
}
\description{
This function summarizes the data in a dataframe according to user-defined groups and functions. In particular, it allows you to calculate the intensity and frequency of disease attack per plot.
}
