% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{resume_pivot_wider}
\alias{resume_pivot_wider}
\title{Pivot wider a resume dataframe}
\usage{
resume_pivot_wider(df_resume)
}
\arguments{
\item{df_resume}{a dataframe returned by resume_data function}
}
\value{
a dataframe in wider format
}
\description{
Pivot wider a resume dataframe
}
