% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{severity_diseased}
\alias{severity_diseased}
\title{calculation of disease severity}
\usage{
severity_diseased(vecteur)
}
\arguments{
\item{vecteur}{a numeric vector of disease severity observations in percentage}
}
\value{
a numeric value of intensity
}
\description{
This function calculate the disease severity, as the mean value of severity measurements for diseased units ONLY (leaves, bunches, plants...).
The disease severity is the area of a sampling unit affected by the disease, expressed as a percentage of the total area.
}
\examples{
vec <- c(100, 20, 30, 0, 5, 0, 0)
severity_diseased(vec)
}
