% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardise.R
\name{standardise_data}
\alias{standardise_data}
\title{Standardize an experimental data file or data frame}
\usage{
standardise_data(
  input_data,
  lookup_table = c(Mildiou_Feuille = "PM_LEAF_PC", Mildiou_Grappe = "PM_BER_PC",
    Oidium_Grappe = "UN_BER_PC", Oidium_Feuille = "UN_LEAF_PC", Stade.phénologique =
    "bbch_stage", Stade.phenologique = "bbch_stage", `Stade phenologique` = "bbch_stage",
    Placette = "plot_id", Bloc = "block_id", Date = "observation_date", Code.essai =
    "experiment_id")
)
}
\arguments{
\item{input_data}{A path to a .csv or .xlsx file to import, or a data.frame already loaded in R.}

\item{lookup_table}{Optional. A named vector mapping original column names to standardized names. If NULL, a default mapping is used.}
}
\value{
A cleaned and structured data.frame, ready for use in downstream processing and analysis.
}
\description{
Standardize an experimental data file or data frame
}
