% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{standardise_pom_csv}
\alias{standardise_pom_csv}
\title{Standardisation of POM csv file}
\usage{
standardise_pom_csv(file, skip_forecast = TRUE)
}
\arguments{
\item{file}{Path to the data leading to a csv file exported from Epicure (IFV Information system)}

\item{skip_forecast}{Logical. If \code{TRUE} (default), the first 14 data rows are removed before processing (commonly forecast values).}
}
\value{
A standardized data.frame with consistent column names and structure.
}
\description{
Standardize a raw POM (EPICURE) csv file
}
