% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startbox-package.R
\docType{package}
\name{startbox-package}
\alias{startbox-package}
\title{startbox: visualisation and standardization of experimental data in plant protection}
\description{
This package is used for management and visualisation of experimental data in plant protection trials.
}
\author{
\strong{Maintainer}: Xavier Delpuech \email{xavier.delpuech@vignevin.com}

Authors:
\itemize{
  \item Hervé Maire \email{herve.maire@vignevin.com}
  \item Anne-Sophie Chazalmartin \email{anne-sophie.chazalmartin@vignevin.com}
}

Other contributors:
\itemize{
  \item  Institut Français de la Vigne et du Vin [copyright holder]
}

}
\keyword{internal}
