% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startbox-package.R
\docType{package}
\name{startbox}
\alias{startbox}
\alias{startbox-package}
\title{startbox}
\description{
This package is used to manage and display experimental plant protection data.
}
\details{
To get started, see the vignette: \code{vignette("startbox")}.
}
\seealso{
\link{vignette("startbox")}
}
\author{
\strong{Maintainer}: Xavier Delpuech \email{xavier.delpuech@vignevin.com}

Authors:
\itemize{
  \item Hervé Maire \email{herve.maire@vignevin.com}
  \item Anne-Sophie Chazalmartin \email{anne-sophie.chazalmartin@vignevin.com}
}

Other contributors:
\itemize{
  \item  Institut Français de la Vigne et du Vin [copyright holder]
}

}
