% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_on_data}
\alias{stat_on_data}
\title{Statistic pipeline for data analysis}
\usage{
stat_on_data(
  data,
  formula_str,
  alpha = 0.05,
  force_test = NULL,
  p.adj = "bonferroni"
)
}
\arguments{
\item{data}{a data.frame with data}

\item{formula_str}{a string with model formula, as "\verb{FA UN_BER_PC} ~ xp_trt_code"}

\item{alpha}{alpha value for statistical tests}

\item{force_test}{NULL, "anova" or "kruskal". By default set to NULL, and if normality or heteroscedasticity are correct, an Anova test is done, otherwise a Kruskal test is done.}

\item{p.adj}{Method for adjusting p values. see agricolae::LSD.test()}
}
\value{
a list
}
\description{
This function applies a linear model to a dataset to test the effect of one or more factors.
If the conditions of normality and heterosceadiscity are met, an ANOVA test is performed.
Otherwise, a non-parametric Kruskal-Walis test is performed.
}
