% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{statsnames}
\alias{statsnames}
\title{Names of stats available in an user_data objetc}
\usage{
statsnames(self)
}
\arguments{
\item{self}{self an instance of the \code{UserData} R6 class}
}
\value{
a vector of names of prepared data
}
\description{
Names of stats available in an user_data objetc
}
