% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_posix}
\alias{to_posix}
\title{Convert heterogeneous date/time inputs to POSIXct (UTC)}
\usage{
to_posix(v)
}
\arguments{
\item{v}{A vector of dates or date-time values. Can be of class
\code{POSIXct}, \code{Date}, or character.}
}
\value{
A vector of class \code{POSIXct} with time zone set to \code{UTC}.
Unparsable values are returned as \code{NA}.
}
\description{
Converts a vector of dates or date-time representations into a
\code{POSIXct} object in the \code{UTC} time zone. The function accepts
\code{POSIXct}, \code{Date}, or character inputs and attempts to parse
character values using a predefined set of common date and date-time
formats. Values that cannot be parsed are returned as \code{NA}.
}
\details{
If \code{v} is already a \code{POSIXct} object, it is returned unchanged.
If \code{v} is a \code{Date} object, it is converted to \code{POSIXct}
with an implicit time of midnight (00:00:00) in UTC.
Character inputs are parsed iteratively using several common
date and date-time formats until all possible values are converted.
}
\examples{
\dontrun{
to_posix("2024-03-15 14:30")
to_posix(c("15/03/2024", "2024-03-16 08:00"))
to_posix(as.Date("2024-03-15"))
}
}
