% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_excel.R
\name{wrapper_data}
\alias{wrapper_data}
\title{Wrapper to import and export observation data in Excel}
\usage{
wrapper_data(self)
}
\arguments{
\item{self}{A \code{user_data} object (R6 class) containing the path to the Excel file.}
}
\value{
Invisibly returns the path to the new Excel file generated with additional data sheets.
}
\description{
This function first imports all \verb{data_*} sheets from the Excel file into the \code{obs_data}
attribute of the R6 object, then exports them into a new Excel file with versioning.
}
