% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metrics.R
\name{efficacy}
\alias{efficacy}
\title{calcul efficacy}
\usage{
efficacy(value, value_tnt)
}
\arguments{
\item{value}{numeric value}

\item{value_tnt}{numeric value of TNT}
}
\description{
calcul efficacy
}
