% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{efficacy}
\alias{efficacy}
\title{calcul efficacy}
\usage{
efficacy(value, value_tnt)
}
\arguments{
\item{value}{numeric value}

\item{value_tnt}{numeric value of TNT}
}
\value{
a vector of calculated efficacy
}
\description{
This function calculate the treatment efficacy.
Efficacy is the level of reduction of target harmful organisms or damage caused by them to the plant,
after application of a treatment, compared to an untreated control.
Efficacy is expressed in percentage.
}
