% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_excel.R
\name{get_template_excel}
\alias{get_template_excel}
\title{get_template_excel}
\usage{
get_template_excel(destination_path = NULL)
}
\arguments{
\item{destination_path}{the file path where the template will be saved}
}
\description{
get_template_excel
}
