% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metrics.R
\name{intensity}
\alias{intensity}
\title{calculation of disease intensity}
\usage{
intensity(vecteur)
}
\arguments{
\item{vecteur}{a numeric vector of disease severity observations in percentage}
}
\value{
a numeric value of intensity
}
\description{
This function calculate the disease intensity, as the mean value of severity measurements.
The disease intensity is the amount of disease present in the population,
expressed in percentage.
}
\examples{
vec <- c(100, 20, 30, 0, 5, 0, 0)
intensity(vec)
}
