% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpbar}
\alias{plot_xpbar}
\title{Barplot summary for an experiment}
\usage{
plot_xpbar(
  self,
  stats = NULL,
  calculation_choices = NULL,
  one_plot = TRUE,
  text_stat = TRUE,
  max_y = NULL,
  bar_color = NULL,
  bar_width = 0.7,
  border_tnt = FALSE,
  show_tnt = TRUE,
  code_tnt = "TNT",
  show_errorbar = TRUE,
  show_data = TRUE,
  short_names = TRUE,
  factor_choice = NULL,
  ...
)
}
\arguments{
\item{self}{An instance of the \code{UserData} R6 class}

\item{stats}{character, the name of a stats list inside self$stats.}

\item{calculation_choices}{vector of character, choice(s) of calculation to plot. For example c("FA UN_BER_PC","IA UN_BER_PC")}

\item{one_plot}{boolean. If TRUE, all calculations are plot on an unique plot using fill argument, else each calculation is on a independent plot using factet}

\item{text_stat}{boolean. If TRUE, an annotation is added to the plot with the test done and its p-value}

\item{max_y}{Upper limit of the y axis. If NULL, automatic adjustment.}

\item{bar_color}{Bar fill colour.}

\item{bar_width}{bar width}

\item{border_tnt}{if true, the border color is set as red for TNT}

\item{show_tnt}{if false, TNT are removed from the graph}

\item{code_tnt}{a string to identify in TNT in the row of the dataframe by default "TNT"}

\item{show_errorbar}{display of error bars}

\item{show_data}{display data points}

\item{short_names}{to have shorter names for fill removing the last part of the string after the last space occurence}

\item{factor_choice}{character. Name of the factor to plot. If NULL only the first factor is used}

\item{...}{other parameters for labs (title, x, y,fill)}
}
\value{
a barplot
}
\description{
This function automatically generates a bar graph from observation or experimental data.
}
\examples{
# example code

\dontrun{

# Before creating the graph, it is necessary to have compiled the statistics.
prepare_data(mydata, df = "data_G1")
test_stats(mydata, prep_data = "prp1_data_G1")

# 3. Generating the barplot
plot_xpbar(mydata, stats = "prp1_data_G1")

# Custom example with data points and error bars
plot_xpbar(mydata,
           stats = "prp1_data_G1",
           show_data = TRUE,
           show_errorbar = TRUE,
           bar_color = c("#d9f0d3", "#1b7837"),
           )
}

}
