% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpbox}
\alias{plot_xpbox}
\title{Boxplot summary for an experiment}
\usage{
plot_xpbox(
  data,
  titre = NULL,
  echelle = NULL,
  caption = "IFV+",
  show_dots = FALSE
)
}
\arguments{
\item{data}{Raw data containing at least plot_id and a "_PC" variable.}

\item{titre}{Main title of the graph (optional).}

\item{echelle}{Upper limit of the y-axis (optional, otherwise automatic).}

\item{caption}{A string used as the caption text displayed at the bottom of the plot. Default is "IFV+".}
}
\value{
A \code{ggplot2} object.
}
\description{
This function automatically generates a box plot
}
