% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpbox}
\alias{plot_xpbox}
\title{Boxplot summary for an experiment}
\usage{
plot_xpbox(
  data,
  echelle = NULL,
  show_dots = FALSE,
  calculation_type = "frequence",
  ...
)
}
\arguments{
\item{data}{Raw data containing at least plot_id, xp_trt_code and a "_PC" variable.}

\item{echelle}{Upper limit of the y-axis (optional, otherwise automatic).}

\item{...}{other parameters for labs (title, x, y,fill)}
}
\value{
A \code{ggplot2} object.
}
\description{
This function automatically generates a box plot
}
