% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_xpheat}
\alias{plot_xpheat}
\title{Plot a Heatmap of Experimental Data}
\usage{
plot_xpheat(
  data,
  variable,
  titre = NULL,
  echelle = NULL,
  orientation = c("vertical", "horizontal"),
  residus = FALSE,
  fill = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{A dataframe containing at least 'plot_id', 'plot_x', 'plot_y', the variable of interest, and optionally 'calculation'.}

\item{variable}{Character. The name of the column representing the variable to display (e.g., "PM_LEAF_PC").}

\item{titre}{(optional) Character. Title of the plot. If NULL, a default title based on the variable is used.}

\item{echelle}{(optional) Numeric. Maximum value for the fill color scale. If NULL, an automatic scale is calculated.}

\item{orientation}{Character. Either "vertical" or "horizontal" to control plot orientation. Default is "vertical".}

\item{residus}{Logical. If TRUE, centers the values by subtracting the global mean (residual visualization).}

\item{fill}{Optional. A vector of two colors (e.g., c("yellow", "red")) to manually define the fill gradient. Ignored if \code{residus = TRUE}.}

\item{caption}{Character. A string used as the caption text displayed at the bottom of the plot. Default is "IFV+".}
}
\value{
A ggplot2 heatmap object.
}
\description{
Generates a heatmap based on experimental observation data,
using \code{plot_x} and \code{plot_y} coordinates and coloring according to a selected disease variable (_PC columns).
}
\details{
\itemize{
\item Aggregates values by \code{plot_id}, taking the mean when multiple observations exist.
\item If no *_PC variable is found in the data, the function will stop with an error.
\item Useful for visualizing spatial patterns of disease in experimental trials.
}
}
