% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{resume_data}
\alias{resume_data}
\title{Title}
\usage{
resume_data(
  data,
  var_col,
  group_cols,
  funs = list(intensite = mean, frequence = frequency)
)
}
\arguments{
\item{data}{a dataframe to resume}

\item{var_col}{character, the colname of the variable to plot}

\item{group_cols}{the colname of the group}

\item{funs}{statistic to plot, a vector of one or two statistic. by default c("mean","frequency")}
}
\value{
a dataframe ready for plotting
}
\description{
Title
}
