% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_user_data.R
\docType{class}
\name{user_data}
\alias{user_data}
\title{user_data}
\description{
\code{user_data} encapsulates the data and metadata associated with a trial.
}
\details{
An R6 class for managing trial data

This class uses the R6 paradigm (mutable objects).
}
\section{Fields}{

\describe{
\item{name}{\code{character} name of the user_data object, by default "user_data".}
\item{excel_data_trial}{\code{character} path to the current Excel file used for import/export operations.}
\item{obs_data}{A named \code{list} of data.frames, each corresponding to a sheet or file containing observation data}
\item{metadata}{A \code{list} storing metadata tables, such as "placette", "modalite"}
\item{plot_tnt_association}{A \code{list} storing association tables between plot and TNT}
\item{prepared_data}{A \code{list} storing prepared_data tables ready for stats analysis as produced by the function prepare_data()}
\item{stats}{A \code{list} storing stats results as results of the function stats_tests()}
\item{traceability}{A \code{data.frame} storing a log of all operations performed on the data (import, export, update, etc.).}
\item{dictionary}{A \code{data.frame} storing the data dictionary}
\item{meteo}{A \code{data.frame} storing the meteo data}
}
}

\section{Methods}{

\describe{
\item{initialize()}{Initializes a new \code{user_data} object}
\item{add_metadata()}{Adds or updates a metadata element to the \code{user_data} object}
\item{add_obs()}{Adds or updates a data element to the \code{user_data} object}
\item{show_obs_data()}{Displays the first few rows of each observation dataset stored in the \code{user_data} object}
\item{log_trace()}{Adds a new entry to the \code{traceability} log stored in the \code{user_data} object}
}
}

\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name of the user_data object, by default "user_data".}

\item{\code{excel_data_trial}}{Path to the current Excel file used for import/export operations.}

\item{\code{obs_data}}{A named list of data.frames, each corresponding to a sheet or file containing observation data.}

\item{\code{metadata}}{A list storing metadata tables, such as "placette" and "modalite".}

\item{\code{plot_tnt_association}}{A list storing association tables between plot and TNT.}

\item{\code{prepared_data}}{A list storing prepared_data tables ready for stats analysis as produced by the function prepare_data()}

\item{\code{stats}}{A list storing stats results as results of the function stats_tests()}

\item{\code{traceability}}{A data.frame storing a log of all operations performed on the data (import, export, update, etc.).}

\item{\code{dictionary}}{A data.frame storing the data dictionary.}

\item{\code{meteo}}{A data.frame storing the meteo data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserData-new}{\code{user_data$new()}}
\item \href{#method-UserData-add_metadata}{\code{user_data$add_metadata()}}
\item \href{#method-UserData-add_obs}{\code{user_data$add_obs()}}
\item \href{#method-UserData-show_obs_data}{\code{user_data$show_obs_data()}}
\item \href{#method-UserData-log_trace}{\code{user_data$log_trace()}}
\item \href{#method-UserData-clone}{\code{user_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new \code{user_data} object. If no Excel file is provided, a default template is used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$new(trial_file = NULL, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trial_file}}{Optional. Path to the Excel file to load. If NULL, a default template is used.}

\item{\code{name}}{optional. A string to name the new \code{user_data} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the \code{UserData} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Adds or updates a metadata element in the \code{metadata} slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$add_metadata(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A single character string specifying the name of the element to add.}

\item{\code{value}}{The object to add (e.g., a data.frame, list, or other metadata).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None. Modifies the \code{metadata} list in-place.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-add_obs"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-add_obs}{}}}
\subsection{Method \code{add_obs()}}{
Adds a new observation dataset to \code{obs_data} or replaces an existing one.
Automatically logs the operation and adds provenance columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$add_obs(name, df, source_file = NULL, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the observation source.}

\item{\code{df}}{A data.frame containing the observation data.}

\item{\code{source_file}}{the source file path}

\item{\code{overwrite}}{Logical. If TRUE, replaces an existing entry with the same name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None. The object is modified in place.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-show_obs_data"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-show_obs_data}{}}}
\subsection{Method \code{show_obs_data()}}{
Displays the first few rows of each observation dataset stored in \code{obs_data}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$show_obs_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of data.frames (head of each dataset).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-log_trace"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-log_trace}{}}}
\subsection{Method \code{log_trace()}}{
This function adds a new entry to the \code{traceability} log stored in the R6 object.
It records the type of operation, the target file or sheet name(s), and the timestamp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$log_trace(operation, filename, description = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{operation}}{A character string describing the action performed (e.g. \code{"import"}, \code{"export"}, \code{"update"}).}

\item{\code{filename}}{A character string indicating the name(s) of the file or sheet involved in the operation.}

\item{\code{description}}{A description of the operation to store in log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. This function updates the internal \code{traceability} data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
