% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_user_data.R
\name{user_data}
\alias{user_data}
\title{User Data R6 Class}
\value{
a  "UserData" object

updated UserData
}
\description{
User Data R6 Class

User Data R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserData-new}{\code{user_data$new()}}
\item \href{#method-UserData-add_metadata}{\code{user_data$add_metadata()}}
\item \href{#method-UserData-add_obs}{\code{user_data$add_obs()}}
\item \href{#method-UserData-show_obs_data}{\code{user_data$show_obs_data()}}
\item \href{#method-UserData-combine_data_obs}{\code{user_data$combine_data_obs()}}
\item \href{#method-UserData-clone}{\code{user_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-new}{}}}
\subsection{Method \code{new()}}{
create a new "UserData" object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$new(excel_model = NULL, excel_data_trial = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{excel_model}}{excel template path}

\item{\code{excel_data_trial}}{excel data trial path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-add_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-add_metadata}{}}}
\subsection{Method \code{add_metadata()}}{
Adds an element to the \code{metadata} list stored in the R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$add_metadata(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A single character string specifying the name of the element to add.}

\item{\code{value}}{The object to add (e.g., a data.frame, list, or other metadata).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None. Modifies the \code{metadata} list in-place.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-add_obs"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-add_obs}{}}}
\subsection{Method \code{add_obs()}}{
Internal function to add or update a dataset of observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$add_obs(name, df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the observation to add}

\item{\code{df}}{dataframe with observation data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-show_obs_data"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-show_obs_data}{}}}
\subsection{Method \code{show_obs_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$show_obs_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-combine_data_obs"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-combine_data_obs}{}}}
\subsection{Method \code{combine_data_obs()}}{
Combines all loaded observation datasets, applies type harmonization, and merges them with an existing Excel trial file.
If the Excel trial file does not exist, it is created from a template. Also reads available metadata sheets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$combine_data_obs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A combined \code{data.frame} of all observations, stored in \code{self$combined_data} and returned invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
