% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_excel.R
\name{wrapper_data}
\alias{wrapper_data}
\title{Wrapper to load all data to do graphs}
\usage{
wrapper_data(self)
}
\arguments{
\item{self}{A \code{user_data} object (R6 class) containing the path to the Excel file.}
}
\value{
Invisibly returns invisible return, but file data loaded
}
\description{
This function first imports all \verb{data_*} sheets from the Excel file into the \code{obs_data}
Loads the metadata if the plot and method sheets are filled in.
}
